+++ base/show.jl
@@ -482,17 +482,17 @@ function _show_default(io::IO, @nospecialize(x))
     print(io,')')
 end
 
-get_active_module() = isdefined(Base, :active_repl) && isdefined(Base.active_repl, :mistate) ?
-    Base.active_repl.mistate.active_module :
-    Main
+active_module() = isdefined(Base, :active_repl) && isdefined(Base.active_repl, :mistate) ?
+                      Base.active_repl.mistate.active_module :
+                      Main
 
 # Check if a particular symbol is exported from a standard library module
 function is_exported_from_stdlib(name::Symbol, mod::Module)
     !isdefined(mod, name) && return false
     orig = getfield(mod, name)
     while !(mod === Base || mod === Core)
+        activemod = active_module()
         parent = parentmodule(mod)
-        activemod = get_active_module()
         if mod === activemod || mod === parent || parent === activemod
             return false
         end
@@ -708,7 +708,7 @@ function show_typealias(io::IO, name::GlobalRef, x::Type, env::SimpleVector, whe
         # Print module prefix unless alias is visible from module passed to
         # IOContext. If :module is not set, default to Main (or current active module).
         # nothing can be used to force printing prefix.
-        from = get(io, :module, get_active_module())
+        from = get(io, :module, active_module())
         if (from === nothing || !isvisible(name.name, name.mod, from))
             show(io, name.mod)
             print(io, ".")
@@ -1024,7 +1024,7 @@ function show_type_name(io::IO, tn::Core.TypeName)
         # Print module prefix unless type is visible from module passed to
         # IOContext If :module is not set, default to Main (or current active module).
         # nothing can be used to force printing prefix
-        from = get(io, :module, get_active_module())
+        from = get(io, :module, active_module())
         if isdefined(tn, :module) && (from === nothing || !isvisible(sym, tn.module, from))
             show(io, tn.module)
             print(io, ".")
@@ -2778,9 +2778,9 @@ MyStruct
 ```
 """
 function dump(arg; maxdepth=DUMP_DEFAULT_MAXDEPTH)
-    # this is typically used interactively, so default to being in Main
-    mod = get(stdout, :module, get_active_module())
-    dump(IOContext(stdout, :limit => true, :module => mod), arg; maxdepth=maxdepth)
+    # this is typically used interactively, so default to being in Main (or current active module)
+    mod = get(stdout, :module, active_module())
+    dump(IOContext(stdout::IO, :limit => true, :module => mod), arg; maxdepth=maxdepth)
 end